// frontend/web/src/services/API/proserve-wb-packaging-api/models/GetComponentResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Component } from './Component';
import {
    ComponentFromJSON,
    ComponentFromJSONTyped,
    ComponentToJSON,
} from './Component';
import type { ComponentMetadata } from './ComponentMetadata';
import {
    ComponentMetadataFromJSON,
    ComponentMetadataFromJSONTyped,
    ComponentMetadataToJSON,
} from './ComponentMetadata';

/**
 * 
 * @export
 * @interface GetComponentResponse
 */
export interface GetComponentResponse {
    /**
     * 
     * @type {Component}
     * @memberof GetComponentResponse
     */
    component: Component;
    /**
     * 
     * @type {ComponentMetadata}
     * @memberof GetComponentResponse
     */
    metadata?: ComponentMetadata;
}

/**
 * Check if a given object implements the GetComponentResponse interface.
 */
export function instanceOfGetComponentResponse(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "component" in value;

    return isInstance;
}

export function GetComponentResponseFromJSON(json: any): GetComponentResponse {
    return GetComponentResponseFromJSONTyped(json, false);
}

export function GetComponentResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): GetComponentResponse {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'component': ComponentFromJSON(json['component']),
        'metadata': !exists(json, 'metadata') ? undefined : ComponentMetadataFromJSON(json['metadata']),
    };
}

export function GetComponentResponseToJSON(value?: GetComponentResponse | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'component': ComponentToJSON(value.component),
        'metadata': ComponentMetadataToJSON(value.metadata),
    };
}