// frontend/web/src/services/API/proserve-wb-packaging-api/models/UpdateComponentRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Workbench Packaging API
 * API in charge of packaging workbench products
 *
 * The version of the OpenAPI document: 2023-10-31T12:00:00Z
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateComponentRequest
 */
export interface UpdateComponentRequest {
    /**
     * The new description for the component
     * @type {string}
     * @memberof UpdateComponentRequest
     */
    componentDescription: string;
}

/**
 * Check if a given object implements the UpdateComponentRequest interface.
 */
export function instanceOfUpdateComponentRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "componentDescription" in value;

    return isInstance;
}

export function UpdateComponentRequestFromJSON(json: any): UpdateComponentRequest {
    return UpdateComponentRequestFromJSONTyped(json, false);
}

export function UpdateComponentRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateComponentRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'componentDescription': json['componentDescription'],
    };
}

export function UpdateComponentRequestToJSON(value?: UpdateComponentRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'componentDescription': value.componentDescription,
    };
}